fix: rename classes to resolve styling conflicts
